use LB google servers, though need to verify that it offers underlying servers round-robin other throttling can happen again; delete unused brightness method
