+++ src/nginx.conf.template
@@ -1,0 +1,95 @@
+worker_processes  8;
+
+error_log  logs/error.log;
+error_log  logs/error.log  notice;
+error_log  logs/error.log  info;
+
+events {
+    worker_connections  1024;
+}
+
+http {
+    server_names_hash_bucket_size  64;
+
+    include       mime.types;
+    default_type  application/octet-stream;
+
+    keepalive_timeout  65;
+
+    server {
+       listen       443 ssl;
+       server_name  khstorelive.azureedge.net;
+
+       ssl_certificate      cert.pem;
+       ssl_certificate_key  key.pem;
+
+       ssl_ciphers  HIGH:!aNULL:!MD5;
+       ssl_prefer_server_ciphers  on;
+
+       access_log  logs/khstore-access.log;
+
+#tsom_cc_activation_masks#       location ~* /.*?tsom_cc_activation_masks.* {
+#tsom_cc_activation_masks#            return 404;
+#tsom_cc_activation_masks#       }
+
+#coverage_maps#       location ~* /.*?/coverage_maps/.* {
+#coverage_maps#            return 404;
+#coverage_maps#       }
+
+#texture_synthesis_online_map_high_res#       location ~* /.*?texture_synthesis_online_map_high_res.* {
+#texture_synthesis_online_map_high_res#            return 404;
+#texture_synthesis_online_map_high_res#       }
+
+#color_corrected_images#       location ~* /.*?color_corrected_images.* {
+#color_corrected_images#            return 404;
+#color_corrected_images#       }
+
+       location / {
+          proxy_set_header                Host khstorelive.azureedge.net;
+          proxy_pass https://#khstorelive.azureedge.net#;
+          proxy_pass_request_headers      on;
+          proxy_ssl_verify              off;
+       }
+    }
+
+    server {
+       listen       443 ssl;
+       server_name  kh.ssl.ak.tiles.virtualearth.net;
+
+       ssl_certificate      cert.pem;
+       ssl_certificate_key  key.pem;
+
+       ssl_ciphers  HIGH:!aNULL:!MD5;
+       ssl_prefer_server_ciphers  on;
+
+       access_log  logs/virtualearth-access.log;
+
+       location /health {
+          proxy_pass http://localhost:39871/health;
+          proxy_ssl_verify              off;
+          access_log  logs/health-check.log;
+       }
+
+#bing_aerial#       location ~* /tiles/akh.*? {
+#bing_aerial#          proxy_pass http://localhost:39871;
+#bing_aerial#          proxy_pass_request_headers      on;
+#bing_aerial#          proxy_ssl_verify              off;
+#bing_aerial#          access_log  logs/localgoogle-access.log;
+#bing_aerial#       }
+
+#bing_aerial#       location ~* /tiles/mtx.*? {
+#bing_aerial#          proxy_pass http://localhost:39871;
+#bing_aerial#          proxy_pass_request_headers      on;
+#bing_aerial#          proxy_set_header Host kh.ssl.ak.tiles.virtualearth.net;
+#bing_aerial#          proxy_ssl_verify              off;
+#bing_aerial#          access_log  logs/localgoogle-access.log;
+#bing_aerial#       }
+
+       location / {
+          proxy_set_header Host kh.ssl.ak.tiles.virtualearth.net;
+          proxy_pass https://#kh.ssl.ak.tiles.virtualearth.net#;
+          proxy_pass_request_headers      on;
+          proxy_ssl_verify              off;
+       }
+    }
+}
